+++ CHANGELOG.md
@@ -5,7 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
 ## [unreleased]
-## Added
+### Added
 - Link to chanjo2 MANE coverage overview on case page and panel page
 - More SVI recommendation links on the ACMG page
 - ACMG Bayesian foundation point scale after Tavtigian for variant heat profile
@@ -13,6 +13,8 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Variants query backend allows rank_score filtering
 - Added script to tabulate causatives clinical filter rank
 - Do not display inheritance models associated to ORPHA terms on variant page
+### Fixed
+- Make BA1 fully stand-alone to Benign prediction
 
 ## [4.89.2]
 ## Fixed